Updates to RKG ceremony to caveat publication of internal documents (#10)
* Updates to RKG ceremony to caveat publication of internal documents

Original RKG ceremony script & annotated script were classified and marked as "internal". A caveat has been added to the title page, as it is inappropriate to modify the documents themselves.

* pdfs move to asset directory

The amended PDF files moved to the asset/documents directory. urls updated accordingly. index.md file in the root directory amended - this is the only effective index.md file.

trust-root/index.md is not a functional component of the website - nonetheless, updated that for the aesthetics.